# Conflicts:
#	src/common/Map.cpp
#	src/ui/ObstacleGraphicItem.cpp

+++ src/common/SquareCollider.cpp
@@ -8,7 +8,7 @@
 
 #include "../headers/SquareCollider.h"
 
-SquareCollider::SquareCollider(int x, int y, int angle, int sideLength) {
+SquareCollider::SquareCollider(double x, double y, double angle, double sideLength) {
     this->_position = new Position(x, y, angle);
     this->_colliderType = ect_square_collider;
     this->_sideLength = sideLength;
@@ -18,10 +18,10 @@ bool SquareCollider::CheckCollision(ICollider * collider) {
     return false; // TODO
 }
 
-int SquareCollider::GetSideLength() {
+double SquareCollider::GetSideLength() {
     return this->_sideLength;
 }
 
-void SquareCollider::SetSideLength(int SquareCollider) {
+void SquareCollider::SetSideLength(double SquareCollider) {
     this->_sideLength = SquareCollider;
 }
