+++ src/headers/CircleCollider.h
@@ -9,14 +9,16 @@
 #ifndef __CIRCLECOLLIDER_H__
 #define __CIRCLECOLLIDER_H__
 
+#include <cmath>
 #include "../interfaces/ICircleCollider.h"
 #include "Position.h"
+#include "CollisionChecker.h"
 
 class CircleCollider : public ICircleCollider {
     public:
-        CircleCollider(int x, int y, int angle, int radius);
-        int GetRadius();
-        void SetRadius(int radius);
+        CircleCollider(double x, double y, double angle, double radius);
+        double GetRadius();
+        void SetRadius(double radius);
         bool CheckCollision(ICollider * collider);
 };
