+++ src/headers/RectangleCollider.h
@@ -11,14 +11,15 @@
 
 #include "../interfaces/IRectangleCollider.h"
 #include "Position.h"
+#include "CollisionChecker.h"
 
 class RectangleCollider : public IRectangleCollider {
     public:
-        RectangleCollider(int x, int y, int angle, int width, int height);
-        int GetWidth();
-        void SetWidth(int width);
-        int GetHeight();
-        void SetHeight(int height);
+        RectangleCollider(double x, double y, double angle, double width, double height);
+        double GetWidth();
+        void SetWidth(double width);
+        double GetHeight();
+        void SetHeight(double height);
         bool CheckCollision(ICollider * collider);
 };
